refactor: simplify repository impls using cleaner data source contracts
Repositories no longer unwrap embedded DTOs or call domain-mapping
extensions directly; those concerns are now owned by the data sources.
Replace embedded != null null-checks with isNotEmpty() on the plain list.
